Moved lobby classes into specific sub-packages
