+++ app/src/main/java/org/albard/dubito/app/lobby/messages/LobbyUpdatedMessage.java
@@ -2,7 +2,7 @@ package org.albard.dubito.app.lobby.messages;
 
 import java.util.Set;
 
-import org.albard.dubito.app.lobby.Lobby;
+import org.albard.dubito.app.lobby.models.Lobby;
 import org.albard.dubito.app.messaging.messages.GameMessageBase;
 import org.albard.dubito.app.network.PeerId;
