+++ app/src/main/java/org/albard/dubito/app/lobby/messages/UpdateLobbyInfoMessage.java
@@ -2,8 +2,8 @@ package org.albard.dubito.app.lobby.messages;
 
 import java.util.Set;
 
-import org.albard.dubito.app.lobby.LobbyId;
-import org.albard.dubito.app.lobby.LobbyInfo;
+import org.albard.dubito.app.lobby.models.LobbyId;
+import org.albard.dubito.app.lobby.models.LobbyInfo;
 import org.albard.dubito.app.messaging.messages.GameMessageBase;
 import org.albard.dubito.app.network.PeerId;
