+++ app/src/main/java/org/albard/dubito/app/lobby/models/Lobby.java
@@ -1,4 +1,4 @@
-package org.albard.dubito.app.lobby;
+package org.albard.dubito.app.lobby.models;
 
 import java.util.HashSet;
 import java.util.Set;
