+++ app/src/main/java/org/albard/dubito/app/lobby/models/LobbyId.java
@@ -1,4 +1,4 @@
-package org.albard.dubito.app.lobby;
+package org.albard.dubito.app.lobby.models;
 
 import java.util.UUID;
