+++ app/src/test/java/org/albard/dubito/app/LobbyServerManagementTest.java
@@ -5,10 +5,6 @@ import java.time.Duration;
 import java.util.List;
 import java.util.Set;
 
-import org.albard.dubito.app.lobby.Lobby;
-import org.albard.dubito.app.lobby.LobbyId;
-import org.albard.dubito.app.lobby.LobbyInfo;
-import org.albard.dubito.app.lobby.LobbyServer;
 import org.albard.dubito.app.lobby.messages.CreateLobbyFailedMessage;
 import org.albard.dubito.app.lobby.messages.CreateLobbyMessage;
 import org.albard.dubito.app.lobby.messages.LeaveLobbyMessage;
@@ -17,6 +13,10 @@ import org.albard.dubito.app.lobby.messages.LobbyLeavedMessage;
 import org.albard.dubito.app.lobby.messages.LobbyUpdatedMessage;
 import org.albard.dubito.app.lobby.messages.UpdateLobbyFailedMessage;
 import org.albard.dubito.app.lobby.messages.UpdateLobbyInfoMessage;
+import org.albard.dubito.app.lobby.models.Lobby;
+import org.albard.dubito.app.lobby.models.LobbyId;
+import org.albard.dubito.app.lobby.models.LobbyInfo;
+import org.albard.dubito.app.lobby.server.LobbyServer;
 import org.albard.dubito.app.messaging.MessageSerializer;
 import org.albard.dubito.app.messaging.MessengerFactory;
 import org.albard.dubito.app.messaging.messages.ErrorGameMessageBase;
