+++ app/src/test/java/org/albard/dubito/app/LobbyServerParticipantsTest.java
@@ -5,9 +5,6 @@ import java.time.Duration;
 import java.util.List;
 import java.util.Set;
 
-import org.albard.dubito.app.lobby.Lobby;
-import org.albard.dubito.app.lobby.LobbyInfo;
-import org.albard.dubito.app.lobby.LobbyServer;
 import org.albard.dubito.app.lobby.messages.CreateLobbyMessage;
 import org.albard.dubito.app.lobby.messages.JoinLobbyFailedMessage;
 import org.albard.dubito.app.lobby.messages.JoinLobbyMessage;
@@ -15,6 +12,9 @@ import org.albard.dubito.app.lobby.messages.LeaveLobbyMessage;
 import org.albard.dubito.app.lobby.messages.LobbyJoinedMessage;
 import org.albard.dubito.app.lobby.messages.LobbyLeavedMessage;
 import org.albard.dubito.app.lobby.messages.LobbyUpdatedMessage;
+import org.albard.dubito.app.lobby.models.Lobby;
+import org.albard.dubito.app.lobby.models.LobbyInfo;
+import org.albard.dubito.app.lobby.server.LobbyServer;
 import org.albard.dubito.app.messaging.MessageSerializer;
 import org.albard.dubito.app.messaging.MessengerFactory;
 import org.albard.dubito.app.messaging.messages.ErrorGameMessageBase;
